fix gradient issues in Safari
This resolves something mentioned in issue #10

The gradient did not seem to be showing Safari due to some Z-Index issue that was not present in Chrome. It seems that the Z-Index declarations were only there to fix a problem with being able to click on the navigation links which does not seem to occur any longer anyway, so I just removed those styles.